requirements.txt
# core dependencies
pyttsx3
pyaudio
TTS
python-dotenv
SpeechRecognition
# voice recognition dependencies
pocketsphinx
whisper